+++ src/save-structure/game-objects/game-object-behavior/known-behaviors/health.ts
@@ -1,5 +1,3 @@
-import { HealthState } from "../../../const-data";
-
 import { GameObjectBehavior } from "../game-object-behavior";
 import { BehaviorName } from "./types";
 
@@ -7,7 +5,6 @@ export const HealthBehavior: BehaviorName<HealthBehavior> = "Health";
 export interface HealthBehavior extends GameObjectBehavior {
   name: "Health";
   templateData: {
-    CanBeIncapacitated: boolean;
-    State: HealthState;
+    canBeIncapacitated: boolean;
   };
 }